Time: 7 ms (74.21%), Space: 8.2 MB (14.23%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     bool isPalindrome(int x) {
         int m =x;
-       long long reverse = 0;
+       long reverse = 0;
         while(m!=0)
         {
             reverse = reverse*10 +m%10;
